ml-service/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
xgboost
pandas
numpy
joblib
PyGithub